implementacion metrica cantidad de aperturas

+++ src/main/java/ar/edu/utn/dds/k3003/model/Heladera.java
@@ -37,7 +37,7 @@ public class Heladera {
     @Transient
     float temperaturaMax;
     @Column
-    long cantidadAperturas;
+    int cantidadAperturas;
 
     public Heladera(String nombre) {
         this.nombre=nombre;
